Restructured folder structure and tidied up the main menu
